Update to match latest commits
Configuration edits are now localized to the top of the file.
Added explanation for using plain string instead of emoji.
"Occassional failure to remove users from on the way or waiting lists, cause unknown. " is no longer an issue. This was due to removing items from a list while iterating over it, which is not a valid operation. Now that duplicates are no longer possible, we break after removing an element, avoiding the issue.